FastAPI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.15.0
sqlalchemy==2.0.21
mysql-connector-python==8.2.0
typing-extensions==4.8.0
multipart==0.2.4
python-multipart==0.0.6